+++ internal/applets/shellutils/halt/halt.go
@@ -1,0 +1,97 @@
+//
+// mimixbox/internal/applets/shellutils/halt/halt.go
+//
+// Copyright 2021 Naohiro CHIKAMATSU
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package halt
+
+import (
+	"fmt"
+	"os"
+	"syscall"
+
+	mb "github.com/nao1215/mimixbox/internal/lib"
+
+	"github.com/jessevdk/go-flags"
+)
+
+const cmdName string = "halt"
+
+const version = "1.0.0"
+
+var osExit = os.Exit
+
+// Exit code
+const (
+	ExitSuccess int = iota // 0
+	ExitFailuer
+)
+
+type options struct {
+	Version bool `short:"v" long:"version" description:"Show halt command version"`
+}
+
+func Run() (int, error) {
+	var opts options
+	var err error
+
+	if _, err = parseArgs(&opts); err != nil {
+		return ExitFailuer, nil
+	}
+	return halt()
+}
+
+func halt() (int, error) {
+	fmt.Fprintln(os.Stdout, "The system is going down NOW !!")
+	if err := killInitProcess(); err != nil {
+		return ExitFailuer, err
+	}
+	return ExitSuccess, nil
+}
+
+func killInitProcess() error {
+	process, err := os.FindProcess(1)
+	if err != nil {
+		return err
+	}
+	err = process.Signal(syscall.Signal(mb.ConvSignalNameToNum("SIGUSR1")))
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func parseArgs(opts *options) ([]string, error) {
+	p := initParser(opts)
+
+	args, err := p.Parse()
+	if err != nil {
+		return nil, err
+	}
+
+	if opts.Version {
+		mb.ShowVersion(cmdName, version)
+		osExit(ExitSuccess)
+	}
+
+	return args, nil
+}
+
+func initParser(opts *options) *flags.Parser {
+	parser := flags.NewParser(opts, flags.Default)
+	parser.Name = cmdName
+	parser.Usage = "[OPTIONS]"
+
+	return parser
+}