+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. The format 
 ## [0.31.00] - 2021-12-17
 ### Added
  - clear command.
+ - halt command. However, this version can not shutdown system.
 ## [0.30.00] - 2021-12-15
 ### Added
  - chown command.